requirements.txt
# Core Telegram Bot functionality# Core Telegram Bot functionality

python-telegram-bot==21.6python-telegram-bot==21.6



# Data handling and analysis# Data handling and analysis

pandas==2.2.3pandas==2.2.3

numpy==2.1.3numpy==2.1.3

requests==2.32.3requests==2.32.3



# Financial data APIs# Financial data APIs

yfinance==0.2.43yfinance==0.2.43



# Charting and visualization# Charting and visualization

matplotlib>=3.10.0matplotlib>=3.10.0



# Scheduling and utilities# Scheduling and utilities

schedule==1.2.2schedule==1.2.2

python-dotenv==1.0.1python-dotenv==1.0.1

tqdm==4.66.5tqdm==4.66.5



# LLM Integration# LLM Integration

groq>=0.33.0groq>=0.33.0



# RAG/AI functionality (optional)# RAG/AI functionality (optional)

faiss-cpu==1.8.0.post1faiss-cpu==1.8.0.post1

sentence-transformers==3.1.1sentence-transformers==3.1.1



# Web framework (optional)# Web framework (optional)

fastapi==0.115.5fastapi==0.115.5

uvicorn==0.32.0uvicorn==0.32.0
